refactor: 패키지 구조에 맞게 Test 코드 분리
